+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `dsnpStartBlockNumber` to config to allow setConfig to default a different value than `0` for `fromBlock`
 - Support "dsnp-start-block" for subscriptions to start from the `dsnpStartBlockNumber`
 - Readme configs for Rinkeby and Ropsten
+- `getDSNPRegistryUpdateEvents` defaults to dsnp-start-block
 
 ### Changed
 - Removed esmodule build. All imports will fallback to esmodule compatible commonjs modules